Broken image.
The img code wasn't using the correct syntax.

<img src=""/>

Incorrect: file path was "./imgs" actually is, "imgs"

Open bracket on <img and no closing bracket /> at the end of the argument.

Now comes the fun part of figuring out the best way to include the image behind everything else, and also so that it is responsive.